requirements.txt
streamlit==0.88.0
google-generativeai=1.0.0
retrying==1.3.3
